+++ src/main/java/com/ttasjwi/cafe/domain/Article.java
@@ -1,10 +1,11 @@
 package com.ttasjwi.cafe.domain;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.Objects;
 
 public class Article {
 
+    private Long articleId;
     private String title;
     private String content;
     private String writer;
@@ -13,6 +14,14 @@ public class Article {
     public Article(String title, String content) {
         this.title = title;
         this.content = content;
+    }
+
+    public Long getArticleId() {
+        return articleId;
+    }
+
+    public void setArticleId(Long articleId) {
+        this.articleId = articleId;
     }
 
     public String getTitle() {
@@ -48,9 +57,23 @@ public class Article {
     }
 
     @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Article article = (Article) o;
+        return articleId.equals(article.articleId);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(articleId);
+    }
+
+    @Override
     public String toString() {
-        return "{" +
-                "title='" + title + '\'' +
+        return "Article{" +
+                "articleId=" + articleId +
+                ", title='" + title + '\'' +
                 ", writer='" + writer + '\'' +
                 ", regDateTime=" + regDateTime +
                 '}';